Split mixer code into initialisation and runtime to reduce flash space usage.
